requirements.txt
openai~=0.27.2
PyGithub~=1.58
requests~=2.28.2
